pyproject.toml
[project]
name = "derag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.58.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.53.0",
    "supabase>=2.27.2",
]
